requirements.txt
flask
pymongo
pyyaml
requests
discord.py>=2.0.0 